+++ backend/controllers/Wishlist.js
@@ -13,6 +13,10 @@ exports.create=async(req,res)=>{
 exports.getByUserId=async(req,res)=>{
     try {
         const {id}=req.params
+        if (req.params.id != req.user) {
+            res.status(401).json({message:"User not authorized for this action"})
+            return;
+        }
         let skip=0
         let limit=0
 
@@ -27,6 +31,8 @@ exports.getByUserId=async(req,res)=>{
         const result=await Wishlist.find({user:id}).skip(skip).limit(limit).populate({path:"product",populate:['brand']})
         const totalResults=await Wishlist.find({user:id}).countDocuments().exec()
 
+        console.log("user" + req.params.id + " | WishList getByUserId")
+
         res.set("X-Total-Count",totalResults)
         res.status(200).json(result)
     } catch (error) {
@@ -37,7 +43,13 @@ exports.getByUserId=async(req,res)=>{
 exports.updateById=async(req,res)=>{
     try {
         const {id}=req.params
+        if (req.params.id != req.user) {
+            res.status(401).json({message:"User not authorized for this action"})
+            return;
+        }
         const updated=await Wishlist.findByIdAndUpdate(id,req.body,{new:true}).populate("product")
+
+        console.log("user" + req.params.id + " | updateById")
         res.status(200).json(updated)
     } catch (error) {
         console.log(error);
@@ -47,7 +59,13 @@ exports.updateById=async(req,res)=>{
 exports.deleteById=async(req,res)=>{
     try {
         const {id}=req.params
+        if (req.params.id != req.user) {
+            res.status(401).json({message:"User not authorized for this action"})
+            return;
+        }
         const deleted=await Wishlist.findByIdAndDelete(id)
+
+        console.log("user" + req.params.id + " | deleteById")
         return res.status(200).json(deleted)
     } catch (error) {
         console.log(error);
